+++ booby/decoders.py
@@ -14,32 +14,13 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 
-"""The `errors` module contains all exceptions used by Booby."""
 
+class Model(object):
+    def __init__(self, model):
+        self.model = model
 
-class BoobyError(Exception):
-    """Base class for all Booby exceptions."""
+    def __call__(self, value):
+        if value is None:
+            return
 
-    pass
-
-
-class FieldError(BoobyError):
-    """This exception is used as an equivalent to :class:`AttributeError`
-    for :mod:`fields`.
-
-    """
-
-    pass
-
-
-class ValidationError(BoobyError):
-    """This exception should be raised when a `value` doesn't validate.
-    See :mod:`validators`.
-
-    """
-
-    pass
-
-
-class EncodeError(BoobyError):
-    pass
+        return self.model.decode(value)