+++ CHANGES.rst
@@ -12,7 +12,8 @@ Backwards-incompatible
 Highlights
 ^^^^^^^^^^
 
-* The `Model` class now defines a `deserialize` and `serialize` methods with basic serialization support.
+* The `Model` class now defines a `decode` and `encode` methods with serialization/deserialization support.
+* A `Field` now can receive lists of callable objects, `encoders` and `decoders`, to perform serialization/deserialization.
 * Added a `List` field that can be used to create fields containing lists of objects (even models).
 
 0.5.1 (Jan 31, 2014)